TARG2
# Targ 2.0 (Temporary Admin Rights Granter)

This application was created to improve on the original which was coded
by another author in VB .NET 2.5.

What does Targ 2.0 do?
===============

* Allows admins in a domain environment to temporarily grant
administrative rights to users for a specified time.

*Remove users from administrators group.

* Simple text file logging of rights granted and error logs.

Improvements over Targ 1
=================

* Works with client machines connected through VPN. Uses another method
that does not require WMI/RPC.

* Supports French language pc's and can be modified to allow other
languages.
About Author / Development
===================

I have been coding on and off for a while and this is my first real
project that has practical use. I developed this solution for use on my
helpdesk team at work. It has been challenging figuring out how to make
it all work, but I have learned a lot in the process.

Goals of this project
=============

* Most importantly I hope to get constructive feedback and learn from
the community.

* Provide a helpful solution to others.